Fix PR review: FK types, validation, config improvements and dependencies
