+++ requirements.txt
@@ -9,3 +9,5 @@ Pygments==2.19.2
 python-dotenv==1.2.14
 SQLAlchemy==2.0.46
 uvicorn==0.39.0
+qrcode
+pillow
